Remove surface-manager-finalize.service
:Release Notes:
Remove surface-manager-finalize.service.
Set permission of XDG_RUNTIME_DIR in surface-manager.service.

:Detailed Notes:
surface-manager-finalize.service is supposed to set permission of
wayland socket but it has not been working as expected when
surface-manager restarts. It results in WebAppMgr and chrome processes
get crashed as they fails to connect to the wayland display.
This patch sets the permission and ownership of XDG_RUNTIME_DIR in
surface-manager.service. For the wayland socket, surface-manager will
set a proper permission as submitted in http://gpro.lgsvl.com/212377.
surface-manager-finalize.service is no longer needed and thus removed.

:Testing Performed:
Restart surface-manager and check if the wayland socket file has rw
permission for compositor group.

:QA Notes:

:Issues Addressed:
[PLAT-62975] WAM crash after restart LSM

Change-Id: I3c210e49b6340f43a787015466d3494b2f0994c5
Reviewed-on: http://gpro.lgsvl.com/212334
Reviewed-by: Andre Rosa <andre.rosa@lge.com>
Reviewed-by: Jaeyoon Jung <jaeyoon.jung@lge.com>
Tested-by: Jaeyoon Jung <jaeyoon.jung@lge.com>
Reviewed-by: Sukil Hong <sukil.hong@lge.com>