Use the non-root(system) user for unifiedsearch
:Release Notes:
Use the non-root(system) user for unifiedsearch

:Detailed Notes:
To improve security with the DAC,
we are decide to use non-root user for the service execution.

This patch will change the uid of unifiedsearch into 'system'.
And also provides a service to prepare it's working directory.

:Testing Performed:
Tested locally based on #487 official build

:QA Notes:

:Issues Addressed:
[WRO-8118] Apply the DAC into IST owned services

Change-Id: I3b1ee56f182688af58f0bbe3dbf9015250898fe1
Reviewed-on: http://gpro.lge.com/c/webosose/webos-initscripts/+/334157
Tested-by: Sangwoo Kang <sangwoo82.kang@lge.com>
Reviewed-by: Sangwoo Kang <sangwoo82.kang@lge.com>