Fix the prepare-system-dirs service file
:Release Notes:
Fix the prepare-system-dirs service file

:Detailed Notes:
Based on the emulator set-up guide,
there is no problem to run this service.

But if someone set the emulator manually
and forgot about the serial port,
the service fails to run because of no tty.

This is a fail-safe logic.

:Testing Performed:
Build and runtime tested

:QA Notes:

:Issues Addressed:
[WRO-10169] [OSE Emulator] : Searched key app is not displaying in the terminal.

Change-Id: I925f152019f38141905e977524cba9d25419318b
Reviewed-on: http://gpro.lge.com/c/webosose/webos-initscripts/+/334539
Tested-by: JeongBong Seo <jb.seo@lge.com>
Reviewed-by: JeongBong Seo <jb.seo@lge.com>
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Sangwoo Kang <sangwoo82.kang@lge.com>